backend/requirements.txt
flask
openai
PyYAML
requests